Build egg skeleton.
Copy the same structure used for genturfahi and jbogenturfahi, with
chicken-*.scm files related to various Chicken Scheme interfaces,
a real command-line program, even if it does nothing, and support
for embedding the LICENSE in the executable.

This is a lot of files for a program that does nothing, but it's
a structure I'm comfortable working with.